feat: parse Data.weighted_distribution as Decimal instead of int
Replace fabricated ×100 WTD integers (9400, 8500, 7500) in all test
rows with realistic percentages (94.00, 85.00, 75.00), update the
happy-path assertion to compare against Decimal("94.00"), and change
the loader to use Decimal(row["WTD"]) instead of int(row["WTD"]).

+++ backend/market_data/management/commands/load_csv.py
@@ -80,7 +80,7 @@ class Command(BaseCommand):
                         date=datetime.date.fromisoformat(row["DATETIME"]),
                         defaults={
                             "value": value,
-                            "weighted_distribution": int(row["WTD"]),
+                            "weighted_distribution": Decimal(row["WTD"]),
                             "period_weeks": int(match.group(1)),
                         },
                     )
